Update from original

+++ scripts/travis/install-torch.sh
@@ -1,4 +1,4 @@
-#/usr/bin/env bash
+#/bin/bash
 # Copyright (c) 2015-2016, NVIDIA CORPORATION.  All rights reserved.
 
 set -e
@@ -10,17 +10,25 @@ then
     exit 1
 fi
 INSTALL_DIR=$1
+
+if [ -d "$INSTALL_DIR" ] && [ -e "$INSTALL_DIR/install/bin/th" ]; then
+    echo "Using cached build at $INSTALL_DIR ..."
+    exit 0
+fi
+
+rm -rf $INSTALL_DIR
 mkdir -p $INSTALL_DIR
 
 # install Torch7
 # instructions from: http://torch.ch/docs/getting-started.html
-curl -s https://raw.githubusercontent.com/torch/ezinstall/master/install-deps | bash
 git clone https://github.com/torch/distro.git $INSTALL_DIR --recursive
-cd $INSTALL_DIR; ./install.sh -b
+cd $INSTALL_DIR
+./install-deps
+./install.sh -b
 
 # install custom packages
-${INSTALL_DIR}/install/bin/luarocks install image
+${INSTALL_DIR}/install/bin/luarocks install tds
 ${INSTALL_DIR}/install/bin/luarocks install "https://raw.github.com/deepmind/torch-hdf5/master/hdf5-0-0.rockspec"
 ${INSTALL_DIR}/install/bin/luarocks install "https://raw.github.com/Sravan2j/lua-pb/master/lua-pb-scm-0.rockspec"
-${INSTALL_DIR}/install/bin/luarocks install lightningmdb LMDB_INCDIR=/usr/local/include LMDB_LIBDIR=/usr/local/lib
+${INSTALL_DIR}/install/bin/luarocks install lightningmdb LMDB_INCDIR=/usr/include LMDB_LIBDIR=/usr/lib/x86_64-linux-gnu
